Moved enpoints to the codes controller
