+++ backend/src/main/java/ca/bc/gov/app/controller/client/ClientCodesController.java
@@ -1,6 +1,8 @@
 package ca.bc.gov.app.controller.client;
 
 import ca.bc.gov.app.dto.client.CodeNameDto;
+import ca.bc.gov.app.dto.client.DistrictDto;
+import ca.bc.gov.app.dto.client.IdentificationTypeDto;
 import ca.bc.gov.app.service.client.ClientService;
 import io.micrometer.observation.annotation.Observed;
 import java.time.LocalDate;
@@ -48,6 +50,65 @@ public class ClientCodesController {
     log.info("Requesting a list of active client contact type codes from the client service.");
     return clientService
         .listClientContactTypeCodes(LocalDate.now(), page, size);
+  }
+  
+  @GetMapping("/districts")
+  public Flux<CodeNameDto> getActiveDistrictCodes(
+      @RequestParam(value = "page", required = false, defaultValue = "0")
+      Integer page,
+      @RequestParam(value = "size", required = false, defaultValue = "10")
+      Integer size) {
+    log.info("Requesting a list of districts from the client service.");
+    return clientService.getActiveDistrictCodes(page, size);
+  }
+
+  @GetMapping("/districts/{districtCode}")
+  public Mono<DistrictDto> getDistrictByCode(@PathVariable String districtCode) {
+    log.info("Requesting a district by code {} from the client service.", districtCode);
+    return clientService.getDistrictByCode(districtCode);
+  }
+  
+  @GetMapping("/countries")
+  public Flux<CodeNameDto> countries(
+      @RequestParam(value = "page", required = false, defaultValue = "0")
+      Integer page,
+      @RequestParam(value = "size", required = false, defaultValue = "10")
+      Integer size) {
+    log.info("Requesting a list of countries from the client service.");
+    return clientService.listCountries(page, size);
+  }
+
+  @GetMapping("/countries/{countryCode}")
+  public Mono<CodeNameDto> getCountryByCode(
+      @PathVariable String countryCode) {
+    log.info("Requesting a country by code {} from the client service.", countryCode);
+    return clientService.getCountryByCode(countryCode);
+  }
+
+  @GetMapping("/countries/{countryCode}/provinces")
+  public Flux<CodeNameDto> listProvinces(
+      @PathVariable String countryCode,
+      @RequestParam(value = "page", required = false, defaultValue = "0")
+      Integer page,
+      @RequestParam(value = "size", required = false, defaultValue = "10")
+      Integer size) {
+    log.info("Requesting a list of provinces by country code {} from the client service.",
+        countryCode);
+    return clientService
+        .listProvinces(countryCode, page, size);
+  }
+  
+  @GetMapping("/identification-types")
+  public Flux<IdentificationTypeDto> identificationTypes() {
+    log.info("Requesting a list of identification type codes.");
+    return clientService.getAllActiveIdentificationTypes(LocalDate.now());
+  }
+
+  @GetMapping("{idCode}")
+  public Mono<CodeNameDto> getIdentificationTypeByCode(
+      @PathVariable String idCode) {
+    log.info("Requesting an identification type by code {}.", idCode);
+    return clientService.getIdentificationTypeByCode(idCode);
   }
 
 }
