+++ frontend/cypress/e2e/FormGeneral.cy.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 describe("General Form", () => {
   beforeEach(() => {
-    cy.intercept("GET", "/api/districts?page=0&size=250", {
+    cy.intercept("GET", "/api/codes/districts?page=0&size=250", {
       fixture: "districts.json",
     }).as("getDistricts");
