+++ frontend/src/components/grouping/StaffLocationGroupComponent.vue
@@ -80,7 +80,7 @@ const updateStateProvince = (value: CodeNameType | undefined, property: string) 
 
 // Province related data
 const provinceUrl = computed(
-  () => `/api/countries/${selectedValue.country.value}/provinces?page=0&size=250`,
+  () => `/api/codes/countries/${selectedValue.country.value}/provinces?page=0&size=250`,
 );
 
 const resetProvinceOnChange = (receivedCountry: any) => {
