+++ frontend/src/pages/FormBCeIDPage.vue
@@ -392,7 +392,7 @@ const scrollToNewContact = () => {
 };
 
 const districtsList = ref([]);
-useFetchTo("/api/districts?page=0&size=250", districtsList);
+useFetchTo("/api/codes/districts?page=0&size=250", districtsList);
 const formattedDistrictsList = computed(() =>
   districtsList.value.map((district) => ({
     ...district,
