+++ frontend/src/pages/staffform/LocationsWizardStep.vue
@@ -68,7 +68,7 @@ const countryList = ref([]);
 
 const fetch = () => {
   if (props.active)
-    useFetchTo("/api/countries?page=0&size=250", countryList);
+    useFetchTo("/api/codes/countries?page=0&size=250", countryList);
 };
 
 watch(() => props.active, fetch);
