+++ frontend/tests/components/components/grouping/StaffLocationGroupComponent.cy.ts
@@ -13,11 +13,11 @@ describe("<StaffLocationGroupComponent />", () => {
   };
 
   beforeEach(() => {
-    cy.intercept("GET", "/api/countries/CA/provinces?page=0&size=250", {
+    cy.intercept("GET", "/api/codes/countries/CA/provinces?page=0&size=250", {
       fixture: "provinces.json",
     }).as("getProvinces");
 
-    cy.intercept("GET", "/api/countries/US/provinces?page=0&size=250", {
+    cy.intercept("GET", "/api/codes/countries/US/provinces?page=0&size=250", {
       fixture: "states.json",
     }).as("getStates");
