+++ frontend/tests/pages/staff/IndividualClientInformationWizardStep.cy.ts
@@ -3,15 +3,15 @@ import type { FormDataDto } from "@/dto/ApplyClientNumberDto";
 
 describe("<individual-client-information-wizard-step />", () => {
   beforeEach(() => {
-    cy.intercept("GET", "/api/countries/CA/provinces?page=0&size=250", {
+    cy.intercept("GET", "/api/codes/countries/CA/provinces?page=0&size=250", {
       fixture: "provinces.json",
     }).as("getProvinces");
 
-    cy.intercept("GET", "/api/countries/US/provinces?page=0&size=250", {
+    cy.intercept("GET", "/api/codes/countries/US/provinces?page=0&size=250", {
       fixture: "states.json",
     }).as("getStates");
 
-    cy.intercept("GET", "/api/identification-types", {
+    cy.intercept("GET", "/api/codes/identification-types", {
       fixture: "identificationTypes.json",
     }).as("getIdentificationTypes");
   });
